video_widget: Use gtk4paintablesink if possible
As it's used by various Gnome apps such as Snapshot and Showtime,
it's shipped by an increasing number of distros. It can serve as
a drop in replacement for the internal sink while providing various
advantages, such as native YCbCr/YUV, rotation and colorimetry
support, resulting in improved performance and quality.

Thus use it if it is present in the system and offers the desired
features.

This is partucularly helpful on phones, where the current pipeline
usually requires several copies for format conversions and 90 degree
camera rotations.