device: Request 30 fps in absence of framerate negotiation
Sources not capable of negotiating framerates, most notably the
libcamera Pipewire plugin, might still be able to honor a value
requested and otherwise usually just ignore it.

Thus setting a value that we can reasonably assume to be supported has
several benefits:
1. It potentially prevents the source from producing much higher fps than
   useful for the usecase, causing resource waste.
2. It decreases the chances of running into fallback code paths which
   are often badly optimized performance-wise. One example is that
   currently the implicitly set framerate values cause regular
   renegotiations (including buffer re-allocations) in pipewiresrc.
   Another is the internal get_target_bitrate(), currently returning
   uint.MAX.
3. If the source can only produce e.g. 15 fps, that's usually handled
   without major issues.

Thus follow the example of Snapshot, which has been doing this for a
while and gets more testing with libcamera, and set framerate to 30/1.